Header options

Properties to modify grid's header, like creating grouped header.

grid_header(
  grid,
  complexColumns = NULL,
  align = NULL,
  valign = NULL,
  height = NULL
)

grid_complex_header(grid, ..., height = 80)

Arguments

grid
A table created with datagrid.

complexColumns
list. This options creates new parent headers of the multiple columns which includes the headers of specified columns, and sets up the hierarchy.

align
Horizontal alignment of the header content. Available values are 'left', 'center', 'right'.

valign
Vertical alignment of the row header content. Available values are 'top', 'middle', 'bottom'.

height
Numeric. The height of the header area.

...
Named arguments to merge columns under a common header, e.g. newcol = c("col1", "col2").

Value

A datagrid htmlwidget.

Examples

library(toastui)

datagrid(rolling_stones_50) %>%
  grid_header(
    align = "left",
    height = "150px"
  )

# Create columns groups
datagrid(iris) %>%
  grid_complex_header(
    "Sepal" = c("Sepal.Length", "Sepal.Width"),
    "Petal" = c("Petal.Length", "Petal.Width")
  )

# or use the full form to use more options
datagrid(iris) %>%
  grid_columns(
    vars = c("Petal.Length", "Petal.Width"),
    header = c("Length", "Width")
  ) %>%
  grid_header(
    complexColumns = list(
      list(
        header = "Sepal",
        name = "Sepal",
        hideChildHeaders = TRUE,
        resizable = TRUE,
        childNames = c("Sepal.Length", "Sepal.Width")
      ),
      list(
        header = "Petal",
        name = "Petal",
        childNames = c("Petal.Length", "Petal.Width")
      )
    ),
    height = 80,
    valign = "middle"
  )