Calendar Week Options

Options for daily, weekly view.

Usage

cal_week_options(
  cal,
  startDayOfWeek = NULL,
  daynames = NULL,
  narrowWeekend = NULL,
  workweek = NULL,
  showNowIndicator = NULL,
  showTimezoneCollapseButton = NULL,
  timezonesCollapsed = NULL,
  hourStart = NULL,
  hourEnd = NULL,
  eventView = TRUE,
  taskView = FALSE,
  collapseDuplicateEvents = NULL,
  ...
)

Arguments

cal

A calendar() object.

startDayOfWeek

Numeric. The start day of week.

daynames

Vector. The day names in weekly and daily. Default values are 'Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'.

narrowWeekend

Logical. Make weekend column narrow(1/2 width).

workweek

Logical. Show only 5 days except for weekend.

showNowIndicator

Display or not the current time indicator in the weekly/daily view.

showTimezoneCollapseButton

Logical. Show a collapse button to close multiple timezones

timezonesCollapsed

Logical. An initial multiple timezones collapsed state.

hourStart

Numeric. Can limit of render hour start.

hourEnd

Numeric. Can limit of render hour end.

eventView

Show the all day and time grid in weekly, daily view. The default value is TRUE. If the value is a vector, it can be "allday", "time".

taskView

Show the milestone and task in weekly, daily view. The default value is FALSE. If the value is a vector, it can be "milestone", "task".

collapseDuplicateEvents

Collapse duplicate events in the daily/weekly view.

...

Additional options.

Value

A calendar htmlwidget.

Note

Online JavaScript documentation: https://github.com/nhn/tui.calendar/blob/main/docs/en/apis/options.md#week

Examples

# Change option for weekly view
calendar(view = "week") %>%
  cal_week_options(
    startDayOfWeek = 1,
    daynames = c("Dim", "Lun", "Mar", "Mer", "Jeu", "Ven", "Sam"),
    narrowWeekend = TRUE
  )