Set global theme options

Properties to customize grid theme, see full list here : https://nhn.github.io/tui.grid/latest/Grid#applyTheme.

set_grid_theme(
  selection.background = NULL,
  selection.border = NULL,
  scrollbar.border = NULL,
  scrollbar.background = NULL,
  scrollbar.emptySpace = NULL,
  scrollbar.thumb = NULL,
  scrollbar.active = NULL,
  outline.border = NULL,
  outline.showVerticalBorder = NULL,
  frozenBorder.border = NULL,
  area.header.border = NULL,
  area.header.background = NULL,
  area.body.background = NULL,
  area.summary.border = NULL,
  area.summary.background = NULL,
  row.even.background = NULL,
  row.even.text = NULL,
  row.odd.background = NULL,
  row.odd.text = NULL,
  row.dummy.background = NULL,
  row.hover.background = NULL,
  cell.normal.background = NULL,
  cell.normal.border = NULL,
  cell.normal.text = NULL,
  cell.normal.showVerticalBorder = NULL,
  cell.normal.showHorizontalBorder = NULL,
  cell.header.background = NULL,
  cell.header.border = NULL,
  cell.header.text = NULL,
  cell.header.showVerticalBorder = NULL,
  cell.header.showHorizontalBorder = NULL,
  cell.rowHeader.background = NULL,
  cell.rowHeader.border = NULL,
  cell.rowHeader.text = NULL,
  cell.rowHeader.showVerticalBorder = NULL,
  cell.rowHeader.showHorizontalBorder = NULL,
  cell.summary.background = NULL,
  cell.summary.border = NULL,
  cell.summary.text = NULL,
  cell.summary.showVerticalBorder = NULL,
  cell.summary.showHorizontalBorder = NULL,
  cell.selectedHeader.background = NULL,
  cell.selectedRowHeader.background = NULL,
  cell.focused.border = NULL,
  cell.focused.background = NULL,
  cell.focusedInactive.border = NULL,
  cell.required.background = NULL,
  cell.required.text = NULL,
  cell.editable.background = NULL,
  cell.editable.text = NULL,
  cell.disabled.background = NULL,
  cell.disabled.text = NULL,
  cell.invalid.background = NULL,
  cell.invalid.text = NULL
)

reset_grid_theme()

Arguments

selection.background
Background color of a selection layer.

selection.border
Border color of a selection layer.

scrollbar.border
Border color of scrollbars.

scrollbar.background
Background color of scrollbars.

scrollbar.emptySpace
Color of extra spaces except scrollbar.

scrollbar.thumb
Color of thumbs in scrollbars.

scrollbar.active
Color of arrows(for IE) or thumb:hover(for other browsers) in scrollbars.

outline.border
Color of the table outline.

outline.showVerticalBorder
Whether vertical outlines of the table are visible.

frozenBorder.border
Border color of a frozen border.

area.header.border
Border color of the header area in the table.

area.header.background
Background color of the header area in the table.

area.body.background
Background color of the body area in the table.

area.summary.border
Border color of the summary area in the table.

area.summary.background
Background color of the summary area in the table.

row.even.background
background color of even row.

row.even.text
text color of even row.

row.odd.background
background color of cells in odd row.

row.odd.text
text color of odd row.

row.dummy.background
background color of dummy row.

row.hover.background
background color of hovered row.

cell.normal.background
Background color of normal cells.

cell.normal.border
Border color of normal cells.

cell.normal.text
Text color of normal cells.

cell.normal.showVerticalBorder
Whether vertical borders of normal cells are visible.

cell.normal.showHorizontalBorder
Whether horizontal borders of normal cells are visible.

cell.header.background
Background color of header cells.

cell.header.border
border color of header cells.

cell.header.text
text color of header cells.

cell.header.showVerticalBorder
Whether vertical borders of header cells are visible.

cell.header.showHorizontalBorder
Whether horizontal borders of header cells are visible.

cell.rowHeader.background
Background color of row's header cells.

cell.rowHeader.border
border color of row's header cells.

cell.rowHeader.text
text color of row's header cells.

cell.rowHeader.showVerticalBorder
Whether vertical borders of row's header cells are visible.

cell.rowHeader.showHorizontalBorder
Whether horizontal borders of row's header cells are visible.

cell.summary.background
Background color of cells in the summary area.

cell.summary.border
border color of cells in the summary area.

cell.summary.text
text color of cells in the summary area.

cell.summary.showVerticalBorder
Whether vertical borders of cells in the summary area are visible.

cell.summary.showHorizontalBorder
Whether horizontal borders of cells in the summary area are visible.

cell.selectedHeader.background
background color of selected header cells.

cell.selectedRowHeader.background
background color of selected row's head cells.

cell.focused.border
border color of a focused cell.

cell.focused.background
background color of a focused cell.

cell.focusedInactive.border
border color of a inactive focus cell.

cell.required.background
background color of required cells.

cell.required.text
text color of required cells.

cell.editable.background
background color of the editable cells.

cell.editable.text
text color of the selected editable cells.

cell.disabled.background
background color of disabled cells.

cell.disabled.text
text color of disabled cells.

cell.invalid.background
background color of invalid cells.

cell.invalid.text
text color of invalid cells.

Value

No return value.

Examples

library(toastui)

# Default is "clean" theme
datagrid(rolling_stones_50)

# others builtins themes
datagrid(rolling_stones_50, theme = "striped")
datagrid(rolling_stones_50, theme = "default")

# Set global theme options
set_grid_theme(
  row.even.background = "#ddebf7",
  cell.normal.border = "#9bc2e6",
  cell.normal.showVerticalBorder = TRUE,
  cell.normal.showHorizontalBorder = TRUE,
  cell.header.background = "#5b9bd5",
  cell.header.text = "#FFF",
  cell.selectedHeader.background = "#013ADF",
  cell.focused.border = "#013ADF"
)

datagrid(rolling_stones_50)

# Remove theme
reset_grid_theme()